NewChat works
I haven't seen any bugs
