+++ .idea/modules.xml
@@ -6,7 +6,6 @@
       <module fileurl="file://$PROJECT_DIR$/RoseChat.iml" filepath="$PROJECT_DIR$/RoseChat.iml" />
       <module fileurl="file://$PROJECT_DIR$/RoseChatApp.iml" filepath="$PROJECT_DIR$/RoseChatApp.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
 </project>
