+++ app/src/main/res/layout/activity_new_chat.xml
@@ -8,30 +8,14 @@
     android:background="@color/colorAccent">
 
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/newChat_fab"
+    <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
-        android:layout_height="360dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="13dp"
-        android:layout_marginEnd="19dp"
-        android:layout_marginRight="19dp"
-        android:visibility="visible"
-        app:backgroundTint="@color/blue_grey_900"
-        app:srcCompat="@android:drawable/ic_input_add" />
-
-<android.support.v7.widget.RecyclerView
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/newChat_recyclerView"
-    android:layout_alignParentTop="true"
-    android:layout_alignParentLeft="true"
-    android:layout_alignParentStart="true"
-    android:layout_above="@+id/newChat_fab">
-
-</android.support.v7.widget.RecyclerView>
+        android:layout_height="match_parent"
+        android:id="@+id/newChat_recyclerView"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true">
+    </android.support.v7.widget.RecyclerView>
