+++ app/src/main/res/layout/new_chat_view.xml
@@ -9,12 +9,15 @@
         android:layout_height="wrap_content"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true">
+        android:layout_alignParentTop="true"
+        android:layout_margin="3dp">
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:paddingTop="5dp"
+            android:paddingBottom="5dp">
 
             <ImageView
                 android:paddingRight="5dp"
@@ -30,10 +33,6 @@
                 android:layout_height="wrap_content"
                 android:textSize="20sp"
                 android:text="Temp_Name" />
-            <CheckBox
-                android:id="@+id/newChat_checkBox"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
         </LinearLayout>
